Move httpAgentStats to HttpClient/middlewares/request.ts
