+++ src/metrics/MetricsAccumulator.ts
@@ -1,7 +1,7 @@
-import { assoc, flatten, map, mapObjIndexed, sum, values } from 'ramda'
+import { assoc, flatten, map, mapObjIndexed, values } from 'ramda'
 import { mean, median, percentile } from 'stats-lite'
 
-import { httpAgent } from '../HttpClient/middlewares/request'
+import { httpAgentStats } from '../HttpClient/middlewares/request'
 import { hrToMillis } from '../utils/time'
 
 export interface Metric {
@@ -47,28 +47,6 @@ function cpuUsage () {
     user: lastCpu.user + diff.user,
   }
   return diff
-}
-
-function countPerOrigin (obj: { [key: string]: any[] }) {
-  try {
-    return mapObjIndexed(val => val.length, obj)
-  } catch (_) {
-    return {}
-  }
-}
-
-export function socketStats () {
-  const socketsPerOrigin = countPerOrigin(httpAgent.sockets)
-  const sockets = sum(values(socketsPerOrigin))
-  const pendingRequestsPerOrigin = countPerOrigin(httpAgent.requests)
-  const pendingRequests = sum(values(pendingRequestsPerOrigin))
-
-  return {
-    pendingRequests,
-    pendingRequestsPerOrigin,
-    sockets,
-    socketsPerOrigin,
-  }
 }
 
 export type StatusTrack = () => EnvMetric[]
@@ -193,8 +171,8 @@ export class MetricsAccumulator {
         production,
       },
       {
-        ...socketStats(),
-        name: 'socket',
+        ...httpAgentStats(),
+        name: 'httpAgent',
         production,
       },
     ]
